# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/localAtendimentos": {
            "get": {
                "tags": [
                    "Local de Atendimentos"
                ],
                "summary": "Listar todos os locais de atendimento",
                "operationId": "e0db85bb813bd5cdadc5b3d5f4f4ff26",
                "responses": {
                    "204": {
                        "description": "Lista de locais de atendimento retornada com sucesso"
                    },
                    "404": {
                        "description": "Nenhum local de atendimento encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Local de Atendimentos"
                ],
                "summary": "Criar um novo local de atendimento",
                "operationId": "4fb570f1aee97e5bb8cee73aec59f34a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "endereco",
                                    "telefone"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Clinica Saúde"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "Rua das Flores, 123"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "(11) 98765-4321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Local de atendimento criado com sucesso"
                    },
                    "404": {
                        "description": "Erro ao criar local de atendimento"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/localAtendimentos/{id}": {
            "put": {
                "tags": [
                    "Local de Atendimentos"
                ],
                "summary": "Atualizar os dados de um Local de Atendimento",
                "operationId": "e1183d1f35a39cf4622b3c2d5430875a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do local de atendimento a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "endereco",
                                    "telefone"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Clinica Saúde"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "Rua das Flores, 123"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "(11) 98765-4321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Local de atendimento atualizado com sucesso"
                    },
                    "404": {
                        "description": "Local de atendimento não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Local de Atendimentos"
                ],
                "summary": "Remover um local de atendimento",
                "operationId": "0884f24243441ae86f2296970b904a62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do local de atendimento a ser removido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Local de atendimento removido com sucesso"
                    },
                    "404": {
                        "description": "Local de atendimento não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}